Issue #136: Added testing framework based on tox and pytest.
Uses tox to run tests and to generate a virtualenv for development.  Added a
file with instructions on use.

Also added a single test, which had previously been part of calc_utils itself.
Many more tests will be needed; this one is included essentially as a
demonstration of the framework.