Issue #136: Added tests for AABB, docstrings, and a minor fix.
Added two tests for the methods in class AABB, and a few more parameter
combinations for the one test for ArcUtils.  Added docstrings to all currently
tested methods.  Tests showed a minor issue in AABB.aabb_in_aabb, where it
could return partially_overlap when it should be returning the stricter
fully_lays_inside if the second AABB shared an edge with the first; this error
was corrected.